requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
pyyaml>=6.0
numpy>=1.24.0
